multiprovider.bbclass: initial checkin
This class helps to select a preferred provider out of a set of package
dependent possible providers. It evaluates the content of two colon
separated lists:

* a package-local list of possible providers; this list can contain
  the special 'none' word

* a distribution-global list of preferred providers; this list can
  contain the special '*' token which will return the first possible
  provider

Its purpose is to avoid dependencies on multiple libraries with
nearly the same functionality by allowing distributions to select
the preferred one.

E.g. people might want a certain package which requires openssl.  But
because other parts of OE (e.g. opkg -> curl) use gnutls, the final
image will contain both the openssl and the gnutls libraries.

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>