autotools: do not output config.log to console
Flooding xterm console output with 10000+ lines of config.log is not
very productive and hides the previously printed information.  Because
the interesting lines are usually in the middle of config.log, there
is not easy way to trim down the output.

Patch copies output of config.log into temp-dir and points to this
file.