gst-plugins-package: set ALLOW_EMPTY_${PN}
Every subpackage has a dependency on '${PN}' because
populate_packages_prepend() specifies "extra_depends=d.expand('${PN}')"
when splitting the subpackages.

Hence, '${PN}' must be always created

Although creation of -plugin-dev and -plugin-staticdev is flawed imo
(.la files are not for development purposes but useless resp. required
for module loading on non GNU/linux platforms, and .a files are not
created for plugins), patch sets ALLOW_EMPTY for -dev and -staticdev
packages for consistency reasons.

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>