Revert "systemd: move the binaries to /sbin/systemd"
This reverts commit d612ca261d12e89e96675c24d9d7456319179720.

This commit causes a lot of mess because keymaps and udev rules are
now located below /sbin/udev.  There are still a lot of places on oe
(and in systemd.bb itself) which refer to the classical /lib/udev
directory (--> 'git grep /lib/udev') and there is an unknown number of
packages whose upstream tarball places udev rules/keymaps in this
directory.

This commit tries to fix one script in one package but breaks systems
silently.  As the 1.4 release is near, possible enhancements of this
patch can not be tested in time anymore and it is better to revert it
completely.

Cc: Ross Burton <ross.burton@intel.com>
Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>