dropbear: hack around different sftp-server path definitions
oe defines libexecdir to be different for each package which breaks
the expectations of dropbear which searches 'sftp-server' within
${libexecdir}.

Patch tries to guess the correct location of the sftp-server helper
program by testing for oe's idea of libexecdir.  People on irc told
they are using the traditional '/usr/libexec' in their distribution so
that this case is handled as well.

Patch is not perfect, but solves the problem for now.

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>