bitbake.conf: fix definition of ${libexecdir}
The GNU standard requires that

  The definition of ‘libexecdir’ is the same for all packages
  [http://www.gnu.org/prep/standards/html_node/Directory-Variables.html]

This means that it must not contain neither:

* multilib qualifiers, nor
* the package name

The recent definition of libexecdir violates this because it can
expand e.g. to

* /usr/lib64/openssh for a 64 bit openssh.bb and
* /usr/lib/dropbear for a 32 bit dropbear.bb

Calls of the 32 bit dropbear to ${libexecdir}/sftp-server will fail
hence.

Patch sets libexecdir to '/usr/lib' as done e.g. by Debian.  Because the
patch changes some basic variables, requires changes of recipes
packaging wildcard '${libexecdir}/*' and can cause silent breakage,
further testing of the generated images is needed and the patch is not
applicable to Dylan.

Wildcard '${libexecdir}/*' in FILES list must be removed after this
change because '${libexecdir}' can become equal to '${libdir}' and these
directories require a different packaging.

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>